feat: api creation for create new vacancy and refactor dto

+++ src/main/java/com/example/candidate_vacancy_management_system/model/Candidate.java
@@ -22,7 +22,7 @@ public class Candidate {
 
     private Date birthdate;
     private String gender;
-    private double currentSalary;
+    private Integer currentSalary;
     private String educationLevel;
     private String schoolName;
 }
